Closes log the versions of dependencies, iina#3490
This commit will:
- Change AppDelegage.applicationWillFinishLaunching to log the IINA version,
  FFmpeg version and the versions of FFmpeg libraries
- Change AppDelegage.applicationDidFinishLaunching to log the mpv version

The version of mpv is not logged in  applicationWillFinishLaunching because
mpv does not provide a static method that returns the version. To obtain
version related information you must construct a mpv object, which has side
effects. So the mpv version is logged in applicationDidFinishLaunching to
preserve the existing order of initialization.