Fix legacy Full Screen is absolutely busted, iina#3543
The commit in the pull request changes the method updateTitle in the class
MainWindowController to use an alternative implementation when the
legacy full screen feature is enabled that avoids calling the method
NSWindow.setTitleWithRepresentedFilename. This works around a defect in
the Cocoa framework that results in IINA crashing with an
NSInvalidArgumentException.