Remove mask window when main window closes, #3558
This is an update to the fix for issue #3558

If the use legacy full screen preference is enabled then when in full-screen
mode on a Mac with a camera housing that intrudes into the screen IINA
displays a black window behind the main window blacks out the screen to the
left and right of the camera. IINA was failing to close that window when
exiting the application while in full-screen mode.

This commit will add code to MainWindowController.windowWillClose to
close the mask window if present.