Update copyright year to 2022
The copyright year displayed in the about window and the macOS "Get Info"
window for the application should reflect the year in which the particular
release of the application was built.

Apparently from a legal perspective updating the copyright year is no longer
required. But users are used to the copyright matching up with the year of the
build and this does provide the user with useful information about the age of
the release. Therefore IINA should continue to update the copyright year.

This commit will:
- Update the copyright year to 2022
- Add a bash script "update_copyright.sh" to the other directory