Backout fix for CPU is consumed by OSC while hidden iina#3601
This commit will remove the changes added to MainWindowController. to
fix the issue with CPU being consumed by the OSC while hidden.

The fix introduced a main thread hang in mpv_get_property called from
PlayerCore.syncUI while running on a timer.

This issue is not critical. Fixing it saves a small amount of energy, so desirable
to eventually fix the issue. Unfortunately I have not been able to reproduce
the hang. Need to figure out how to trigger the problem before proposing
any fix. Backing out faulty fix to stabilize IINA-Plus.