Fix currently open filenames not syncing, iina#3159
This problem is due to a defect in AppKit that has been fixed in macOS
Monterey. The commit changes the window controller to apply a
workaround of reseting the window's title in earlier versions of macOS.
If the title is not reset when a window is reused then the wrong title
will be displayed in the "Window" menu.

The commit will:
- Change MainWindowController.windowWillOpen to set the window's title
  to "Window", the title a NSWindow has by default